backend/requirements-cloud.txt
# 轻量版 requirements（Railway/Render 云部署专用）
# 去掉 chromadb + sentence-transformers，避免超出免费套餐内存
# 部署时请同时设置环境变量 ENABLE_RAG=false

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
anthropic==0.40.0
PyJWT==2.8.0
python-multipart==0.0.6
httpx==0.26.0
